syncdb/src/jp/co/ntt/oss/mapper/TimeDataMapper.java
package jp.co.ntt.oss.mapper;

import java.sql.PreparedStatement;
import java.sql.ResultSet;
import java.sql.SQLException;
import java.sql.Time;
import java.sql.Types;

public class TimeDataMapper implements DataMapper {

	@Override
	public final Object getObject(final ResultSet rset, final int columnIndex)
			throws SQLException {
		return rset.getTime(columnIndex);
	}

	@Override
	public final void setObject(final PreparedStatement pstmt,
			final int parameterIndex, final Object x) throws SQLException {
		pstmt.setTime(parameterIndex, (Time) x);
	}

	public static boolean isTimeType(final int columnType) {
		if (columnType == Types.TIME) {
			return true;
		}

		return false;
	}
}
